fixed `isToggled` param of 'apputils:theme-scrollbars' command
`apputils-extension/src/index.ts` was getting crowded, so I also split out
the theme plugins into their own source file
